+++ lib/gql/relay/index.d.ts
@@ -14,3 +14,4 @@ export * from './with_pagination_filter.mixin';
 export * from './with_timestamps.mixin';
 export * from './with_timestamps_filter.mixin';
 export * from './with_order_by_filter.mixin';
+export * from './relay_edge_connection.builder';